requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
black==23.12.1; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
dj-database-url==2.1.0
dj-rest-auth==5.0.2; python_version >= '3.6'
django==5.0; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
django-environ==0.11.2; python_version >= '3.6' and python_version < '4'
django-filter==23.5; python_version >= '3.7'
django-storages==1.14.2; python_version >= '3.7'
djangorestframework==3.14.0; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
drf-yasg==1.21.7; python_version >= '3.6'
factory-boy==3.3.0; python_version >= '3.7'
faker==22.0.0; python_version >= '3.8'
gunicorn==21.2.0; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
mypy-extensions==1.0.0; python_version >= '3.5'
packaging==23.2; python_version >= '3.7'
pathspec==0.12.1; python_version >= '3.8'
pillow==10.1.0; python_version >= '3.8'
platformdirs==4.1.0; python_version >= '3.8'
psycopg==3.1.16; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0; python_version >= '3.8'
pytz==2023.3.post1
pyyaml==6.0.1; python_version >= '3.6'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.9.0; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
whitenoise==6.6.0; python_version >= '3.8'
